refactor(ors_control_app): extract studio/engine/types and engine/freight
Begin Phase 2 split of the 1,288-line studio/engine.ts. Extract two
self-contained pieces into a new studio/engine/ sub-folder:
- engine/types.ts: TelemetryPoint, TripRecord, GenerationEvent, POI,
  RouteGeometry, FleetMember, GenerationProgress, RouteFetchResult,
  FreightOffer, SnowSqlFn (~125 lines)
- engine/freight.ts: sourceLabelsForRegion, FREIGHT_PRODUCTS,
  generateFreightOffers (synthetic freight offer generator, ~60 lines)

studio/engine.ts now re-exports types and generateFreightOffers from the
new modules, keeping the existing import surface for jobs.ts unchanged.
Lines down 1,288 -> 1,118.

The remaining engine code (routability, POIs, fleet building, telemetry
generation) is more tightly coupled and stays in engine.ts for now.
A follow-up commit can split routability/pois/fleet/telemetry if needed.

Phase 2 of holistic-control-app-refactor-final, partial.
No behaviour change. tsc passes cleanly.
